Category nav bar working
