+++ frontend/components/products/product_nav_index.jsx
@@ -2,6 +2,8 @@ import React from "react";
 import ProductNavShowList from "./product_nav_show_list";
 import {Switch, Route} from "react-router";
 import { Link } from "react-router-dom";
+import ProductIndexItem from "./product_index_item";
+import RecentlyViewItemIndex from "./recently_view_item_index";
 
 
 class ProductNavIndex extends React.Component{
@@ -23,31 +25,30 @@ class ProductNavIndex extends React.Component{
             index === 0 ? word.toLowerCase() : word[0].toUpperCase()+word.slice(1)
             )
         }).join("")
-        const categories = ["Couch and Sofa",
-            "Dining Table",
-            "Kids Bunk Bed",
-            "Accent Chairs",
-            "Dining Chairs",
-            "Outdoor"]
-            // debugger
-        if (categories_selector[currentCategory] === undefined){
+        let index = categories_selector.indexOf(eval(currentCategory))
+        // console.log("couch",couchAndSofa)
+        console.log(currentCategory)
+        console.log("selector",categories_selector)
+        console.log("selector[0]", categories_selector[index])
+
+        // console.log([categories_selector.match(window["couchAndSofa"])] )
+
+        if (categories_selector[index] === undefined){
             return null;
         }
-
         return(
-            <div className="nav-static">
-                <ul>
-                    {
-                    categories_selector[categories_selector.indexOf(currentCategory)].map((products) => {
-                        return(
-                            <ProductNavShowList
-                                key={product.id} products={products}/>
-                        )
-                    })
-                    }
-                </ul>
+            <ul className="nav-static">
+                {
+                categories_selector[index].map((product) => {
 
-            </div>
+                    return(
+                        <ProductIndexItem
+                            product={product}/>
+                    )
+                })
+                }
+
+            </ul>
         )
     }
 }
